fix: no more deadlock
the problem was that np defautls to flaot64 and I use float32 everywhere
else so when concatenating it with the other vectors, the whole thing
becomes np.float64. This would create a 4 bytes offset for each element of
the vector at each reset.
This would destroy the order of elements and fill the pipe up until it
reaches the max size.

+++ src/simulation/src/webots/controllers/controller_vehicle_driver/controller_vehicle_driver.py
@@ -98,7 +98,7 @@ class VehicleDriver(Driver):
 
         if c.camera_horizontal_resolution == 0:
             # empty array
-            camera_data = np.zeros([0])
+            camera_data = np.zeros([0], dtype=np.float32)
         else:
             camera_data = np.array(self.camera.getImageArray(), dtype=np.float32)
             # shape = (camera_horizontal_resolution, 1, 3)
